+++ ItemGenerator/AbilityEffectsCoreData/Guard+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension Guard {
     @NSManaged public var name: String?
     @NSManaged public var procRate: Float
     @NSManaged public var percentGuarded: Float
-    @NSManaged public var amountGuarded: Int16
+    @NSManaged public var amountGuarded: Int32
     @NSManaged public var effectType: String?
 
 }
