+++ ItemGenerator/AbilityEffectsCoreData/LifeGain+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension LifeGain {
         return NSFetchRequest<LifeGain>(entityName: "LifeGain");
     }
 
-    @NSManaged public var amountGain: Int16
+    @NSManaged public var amountGain: Int32
     @NSManaged public var procRate: Float
     @NSManaged public var percentGain: Float
     @NSManaged public var amountPerDamage: Float
